Cleaning. Formatting, and converted comp to lib
