+++ src/lib/Footer.svelte
