+++ src/lib/Post.svelte
